+++ R/create_management_scenario.R
@@ -3,7 +3,7 @@
 #' Defines a management scenario with default values, to be modified by the user
 #'
 #' @param units Number of management units. Alternatively, a data frame with management options (in columns) for a set of units (in rows). Options not specified witl be taken from defaults.
-#' @param annual_demand_by_species A vector or matrix of annual wood demand (m3) by medfate species names. If empty, the scenario is 'bottom-up' (not based on demand).
+#' @param annual_demand_by_species A vector or matrix of annual wood demand (m3) by medfate species names (or groups of species names). If empty, the scenario is 'bottom-up' (not based on demand).
 #'                   If a vector is supplied, the same wood demand is applied for all simulated years. If a matrix is supplied, each row should correspond
 #'                   to a different year.
 #' @param extraction_rate_by_year A vector of extraction rates (%) per year of the simulation, starting at the second year. If specified,
@@ -20,7 +20,7 @@
 #'      \item{\code{scenario_type}: Either 'bottom-up' (no demand is specified), 
 #'            'input_demand' (annual species demand is specified), or 
 #'            'input_rate' when extraction rates are also supplied.}
-#'      \item{\code{annual_demand_by_species}: A vector of annual wood demand (m3) by species 
+#'      \item{\code{annual_demand_by_species}: A vector of annual wood demand (m3) by species (or species groups) 
 #'            (for scenario_type 'bottom-up' or 'input_demand').}
 #'      \item{\code{extraction_rate_by_year}: A vector of extraction rate values per year.}
 #'      \item{\code{units}: A data frame with as many rows as units and management arguments as columns.}
@@ -43,6 +43,10 @@
 #' # and not based on demand
 #' data("defaultPrescriptionsBySpecies")
 #' scen_3 <- create_management_scenario(defaultPrescriptionsBySpecies)
+#' 
+#' # A scenario with three management units and annual demand for one species group 
+#' # and a third species
+#' scen_4 <- create_management_scenario(3,  c("Quercus ilex/Quercus pubescens" = 1000, "Pinus nigra" = 2000))
 #' 
 create_management_scenario<-function(units,
                                      annual_demand_by_species = NULL,