+++ NEWS.md
@@ -3,8 +3,10 @@
 ----------------------------------
 
 
-# Version 2.0.1
+# Version 2.0.2
 - Improvements in fordyn_scenario
+- Allowing management scenarios to specify demand for groups of species
+- Ignoring convex hull checks when calling meteoland interpolation
 
 # Version 2.0.0
 - IMPORTANT: In this version we removed sp-type S4 classes that were deprecated in 1.0.1. Use release version 1.2.1 to continue using deprecated classes.